+++ lib/src/components/form/form.dart
@@ -23,7 +23,7 @@ enum ShadAutovalidateMode {
 /// A typedef representing a map of form field states with dynamic values.
 typedef ShadFormFields =
     Map<
-      Object,
+      String,
       ShadFormBuilderFieldState<ShadFormBuilderField<dynamic>, dynamic>
     >;
 
@@ -61,7 +61,7 @@ class ShadForm extends StatefulWidget {
   final Widget child;
 
   /// Initial values for form fields
-  final Map<Object, dynamic> initialValue;
+  final Map<String, dynamic> initialValue;
 
   /// Whether the form fields are enabled
   final bool enabled;
@@ -99,22 +99,22 @@ class ShadForm extends StatefulWidget {
 class ShadFormState extends State<ShadForm> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final ShadFormFields _fields = {};
-  final Map<Object, dynamic> _value = {};
-  final Map<Object, Function> _transformers = {};
+  final Map<String, dynamic> _value = {};
+  final Map<String, Function> _transformers = {};
   late final ValueNotifier<AutovalidateMode> autovalidateMode;
 
   /// Returns the registered form fields
   ShadFormFields get fields => _fields;
 
   /// Returns the initial form values
-  Map<Object, dynamic> get initialValue => widget.initialValue;
+  Map<String, dynamic> get initialValue => widget.initialValue;
 
   /// Whether the form is enabled
   bool get enabled => widget.enabled;
 
   /// Returns an unmodifiable view of the current form values with
   /// transformations applied
-  Map<Object, dynamic> get value => Map<Object, dynamic>.unmodifiable(
+  Map<String, dynamic> get value => Map<String, dynamic>.unmodifiable(
     _value.map(
       (key, value) =>
           // ignore: avoid_dynamic_calls
@@ -143,7 +143,7 @@ class ShadFormState extends State<ShadForm> {
 
   /// Registers a form field with the specified id
   void registerField(
-    Object id,
+    String id,
     ShadFormBuilderFieldState<ShadFormBuilderField<dynamic>, dynamic> field,
   ) {
     _fields[id] = field;
@@ -153,13 +153,62 @@ class ShadFormState extends State<ShadForm> {
       ..setValue(_value[id]);
   }
 
-  void setInternalFieldValue<T>(Object id, T? value) {
+  /// Sets the value for a form field with the specified id
+  ///
+  /// The [value] parameter is the new value to set for the field.
+  /// If [notifyField] is true (the default), this will call the `didChange`
+  /// method of the field state to update its value and all the side effects,
+  /// like validation and notifying listeners.
+  ///
+  /// If you don't want to trigger those side effects, but only want to change
+  /// the form's map value set [notifyField] to false.
+  void setFieldValue<T>(String id, T? value, {bool notifyField = true}) {
     _value[id] = value;
+    if (notifyField) _fields[id]?.didChange(value);
+  }
+
+  /// Merges the provided entries into the form value; optionally removes keys
+  /// missing from the provided map.
+  ///
+  /// If [notifyFields] is true (the default), this will call the `didChange`
+  /// method of each field state to update its value and all the side effects,
+  /// like validation and notifying listeners.
+  ///
+  /// Only the fields with an updated value will be notified.
+  void setValue(
+    Map<String, dynamic> value, {
+
+    /// When true, notifies the changed form fields of the value changes
+    bool notifyFields = true,
+
+    /// When true, removes keys from the form value that are not present in
+    /// the provided [value] map.
+    bool removeMissing = false,
+
+    /// When true (and `removeMissing` is true), notifies removed fields.
+    bool notifyRemovedFields = false,
+  }) {
+    if (removeMissing) {
+      final keysToRemove = _value.keys
+          .where((key) => !value.containsKey(key))
+          .toList();
+      for (final id in keysToRemove) {
+        removeFieldValue(id, notifyField: notifyRemovedFields);
+      }
+    }
+    for (final entry in value.entries) {
+      final field = _fields[entry.key];
+      final oldValue = _value[entry.key];
+      _value[entry.key] = entry.value;
+      if (notifyFields && field != null && oldValue != entry.value) {
+        field.didChange(entry.value);
+      }
+    }
   }
 
   /// Sets forced internal error for a form field
   /// Throws if the field with [id] is not registered with the form.
-  void setInternalFieldError(Object id, String? error) {
+  void setFieldError(String id, String? error) {
     final field = _fields[id];
     if (field == null) {
       throw FlutterError(
@@ -167,23 +216,30 @@ class ShadFormState extends State<ShadForm> {
         'Make sure the field is registered with the form.',
       );
     }
-    field.setInternalError(error);
+    field.setError(error);
   }
 
   /// Removes internal field value
-  void removeInternalFieldValue(Object id) {
+  ///
+  /// If [notifyField] is true, this will call the form field `didChange` method
+  /// with a `null` value
+  void removeFieldValue(String id, {bool notifyField = false}) {
     _value.remove(id);
+    if (notifyField) {
+      final field = _fields[id];
+      if (field != null) field.didChange(null);
+    }
   }
 
   /// Unregisters a form field
   void unregisterField(
-    Object id,
+    String id,
     ShadFormBuilderFieldState<ShadFormBuilderField<dynamic>, dynamic> field,
   ) {
     _fields.remove(id);
     _transformers.remove(id);
     if (widget.clearValueOnUnregister) {
-      _value.remove(id);
+      removeFieldValue(id);
     }
   }